Moved shaders, textures and geometries
