+++ packages/core/src/Renderer.js
@@ -2,19 +2,19 @@ import AbstractRenderer from './AbstractRenderer';
 import { sayHello } from '@pixi/utils';
 import MaskSystem from './systems/MaskSystem';
 import StencilSystem from './systems/StencilSystem';
-import FilterSystem from './systems/filter/FilterSystem';
+import FilterSystem from './filters/FilterSystem';
 import FramebufferSystem from './systems/FramebufferSystem';
 import RenderTextureSystem from './systems/RenderTextureSystem';
-import TextureSystem from './systems/textures/TextureSystem';
+import TextureSystem from './textures/TextureSystem';
 import ProjectionSystem from './systems/ProjectionSystem';
 import StateSystem from './systems/StateSystem';
-import GeometrySystem from './systems/geometry/GeometrySystem';
-import ShaderSystem from './systems/shader/ShaderSystem';
+import GeometrySystem from './geometry/GeometrySystem';
+import ShaderSystem from './shader/ShaderSystem';
 import ContextSystem from './systems/ContextSystem';
 import BatchSystem from './systems/BatchSystem';
-import TextureGCSystem from './systems/textures/TextureGCSystem';
+import TextureGCSystem from './textures/TextureGCSystem';
 import { RENDERER_TYPE } from '@pixi/constants';
-import UniformGroup from '../shader/UniformGroup';
+import UniformGroup from './shader/UniformGroup';
 import { Matrix } from '@pixi/math';
 import Runner from 'mini-runner';
