+++ packages/core/src/filters/defaultFilter.frag
