+++ packages/core/src/filters/defaultFilter.vert
