+++ packages/core/src/filters/spriteMask/spriteMaskFilter.frag
