+++ packages/core/src/filters/spriteMask/spriteMaskFilter.vert
