+++ packages/core/src/geometry/utils/setVertexAttribArrays.js
