+++ packages/core/src/index.js
@@ -1,11 +1,11 @@
 import * as resources from './textures/resources';
-import * as systems from './renderers/systems';
+import * as systems from './systems';
 
 export { resources };
 export { systems };
 
-export { default as Renderer } from './renderers/Renderer';
-export { default as AbstractRenderer } from './renderers/AbstractRenderer';
+export { default as Renderer } from './Renderer';
+export { default as AbstractRenderer } from './AbstractRenderer';
 export { default as FrameBuffer } from './textures/FrameBuffer';
 export { default as CubeTexture } from './textures/CubeTexture';
 export { default as BaseTexture } from './textures/BaseTexture';
@@ -14,16 +14,16 @@ export { default as TextureMatrix } from './textures/TextureMatrix';
 export { default as RenderTexture } from './textures/RenderTexture';
 export { default as BaseRenderTexture } from './textures/BaseRenderTexture';
 export { default as TextureUvs } from './textures/TextureUvs';
-export { default as State } from './renderers/State';
-export { default as ObjectRenderer } from './renderers/utils/ObjectRenderer';
-export { default as Quad } from './renderers/utils/Quad';
-export { default as QuadUv } from './renderers/utils/QuadUv';
-export { default as checkMaxIfStatmentsInShader } from './renderers/utils/checkMaxIfStatmentsInShader';
+export { default as State } from './State';
+export { default as ObjectRenderer } from './utils/ObjectRenderer';
+export { default as Quad } from './utils/Quad';
+export { default as QuadUv } from './utils/QuadUv';
+export { default as checkMaxIfStatmentsInShader } from './utils/checkMaxIfStatmentsInShader';
 export { default as Shader } from './shader/Shader';
 export { default as Program } from './shader/Program';
 export { default as UniformGroup } from './shader/UniformGroup';
-export { default as SpriteMaskFilter } from './renderers/filters/spriteMask/SpriteMaskFilter';
-export { default as Filter } from './renderers/filters/Filter';
+export { default as SpriteMaskFilter } from './filters/spriteMask/SpriteMaskFilter';
+export { default as Filter } from './filters/Filter';
 export { default as Attribute } from './geometry/Attribute';
 export { default as Buffer } from './geometry/Buffer';
 export { default as Geometry } from './geometry/Geometry';
