+++ packages/core/src/shader/utils/extractUniformsFromSrc.js
