+++ packages/core/src/shader/utils/generateUniformsSync.js
