+++ packages/core/src/shader/utils/setPrecision.js
