+++ packages/core/src/systems/MaskSystem.js
