+++ packages/core/src/systems/StencilSystem.js
