+++ packages/core/src/systems/index.js
@@ -3,4 +3,3 @@
  * @namespace PIXI.systems
  */
 export { default as WebGLSystem } from './WebGLSystem';
-export { default as FilterSystem } from './filter/FilterSystem';
