+++ packages/core/src/utils/QuadUv.js
@@ -1,5 +1,5 @@
-import Geometry from '../../geometry/Geometry';
-import Buffer from '../../geometry/Buffer';
+import Geometry from '../geometry/Geometry';
+import Buffer from '../geometry/Buffer';
 
 /**
  * Helper class to create a quad with uvs like in v4
