+++ packages/core/src/utils/checkMaxIfStatmentsInShader.js
